# public/swagger.json
{
	"components": {
		"examples": {},
		"headers": {},
		"parameters": {},
		"requestBodies": {},
		"responses": {},
		"schemas": {
			"SaveUpdateResADMINwaiter": {
				"properties": {
					"_id": {
						"type": "string"
					},
					"FirstName": {
						"type": "string"
					},
					"LastName": {
						"type": "string"
					},
					"Desgination": {
						"type": "string"
					},
					"cell": {
						"type": "string"
					},
					"JoinDate": {
						"type": "string"
					},
					"adress": {
						"type": "string"
					},
					"createdAt": {
						"type": "string"
					},
					"updatedAt": {
						"type": "string"
					}
				},
				"required": [
					"_id",
					"FirstName",
					"LastName",
					"Desgination",
					"cell",
					"JoinDate",
					"adress",
					"createdAt",
					"updatedAt"
				],
				"type": "object",
				"additionalProperties": false
			},
			"GetAdminwaiter": {
				"properties": {
					"id": {
						"type": "string"
					}
				},
				"required": [
					"id"
				],
				"type": "object",
				"additionalProperties": false
			},
			"SaveReqADMINwaiter": {
				"properties": {
					"FirstName": {
						"type": "string"
					},
					"LastName": {
						"type": "string"
					},
					"Desgination": {
						"type": "string"
					},
					"cell": {
						"type": "string"
					},
					"JoinDate": {
						"type": "string"
					},
					"adress": {
						"type": "string"
					}
				},
				"required": [
					"FirstName",
					"LastName",
					"Desgination",
					"cell",
					"JoinDate",
					"adress"
				],
				"type": "object",
				"additionalProperties": false
			},
			"UpdateReqADMINwaiter": {
				"properties": {
					"_id": {
						"type": "string"
					},
					"FirstName": {
						"type": "string"
					},
					"LastName": {
						"type": "string"
					},
					"Desgination": {
						"type": "string"
					},
					"cell": {
						"type": "string"
					},
					"JoinDate": {
						"type": "string"
					},
					"adress": {
						"type": "string"
					}
				},
				"required": [
					"_id",
					"FirstName",
					"LastName",
					"Desgination",
					"cell",
					"JoinDate",
					"adress"
				],
				"type": "object",
				"additionalProperties": false
			},
			"_LeanDocument__LeanDocument_T__": {
				"properties": {},
				"type": "object"
			},
			"Pick__LeanDocument_T_.Exclude_keyof_LeanDocument_T_.Exclude_keyofDocument._id-or-id-or-__v_-or-%24isSingleNested__": {
				"properties": {
					"_id": {
						"$ref": "#/components/schemas/_LeanDocument__LeanDocument_T__",
						"description": "This documents _id."
					},
					"__v": {
						"description": "This documents __v."
					},
					"id": {
						"description": "The string version of this documents _id."
					}
				},
				"type": "object",
				"description": "From T, pick a set of properties whose keys are in the union K"
			},
			"Omit__LeanDocument_this_.Exclude_keyofDocument._id-or-id-or-__v_-or-%24isSingleNested_": {
				"$ref": "#/components/schemas/Pick__LeanDocument_T_.Exclude_keyof_LeanDocument_T_.Exclude_keyofDocument._id-or-id-or-__v_-or-%24isSingleNested__",
				"description": "Construct a type with the properties of T except for those in type K."
			},
			"LeanDocument_this_": {
				"$ref": "#/components/schemas/Omit__LeanDocument_this_.Exclude_keyofDocument._id-or-id-or-__v_-or-%24isSingleNested_"
			},
			"ADMINwaiter": {
				"$ref": "#/components/schemas/LeanDocument_this_"
			},
			"DeleteAdminwaiter": {
				"properties": {
					"id": {
						"type": "string"
					}
				},
				"required": [
					"id"
				],
				"type": "object",
				"additionalProperties": false
			},
			"SaveUpdateResmenu": {
				"properties": {
					"Iteam_id": {
						"type": "string"
					},
					"IteamName": {
						"type": "string"
					},
					"Price": {
						"type": "number",
						"format": "double"
					},
					"createdAt": {
						"type": "string"
					},
					"updatedAt": {
						"type": "string"
					}
				},
				"required": [
					"Iteam_id",
					"IteamName",
					"Price"
				],
				"type": "object",
				"additionalProperties": false
			},
			"SaveReqmenu": {
				"properties": {
					"IteamName": {
						"type": "string"
					},
					"Price": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"IteamName",
					"Price"
				],
				"type": "object",
				"additionalProperties": false
			},
			"menuadmindoc": {
				"$ref": "#/components/schemas/LeanDocument_this_"
			},
			"Deletemenu": {
				"properties": {
					"id": {
						"type": "string"
					}
				},
				"required": [
					"id"
				],
				"type": "object",
				"additionalProperties": false
			},
			"SaveUpdateReswaiterOder": {
				"properties": {
					"Iteam_id": {
						"type": "string"
					},
					"IteamName": {
						"type": "string"
					},
					"Price": {
						"type": "number",
						"format": "double"
					},
					"password": {
						"type": "string"
					},
					"createdAt": {
						"type": "string"
					},
					"updatedAt": {
						"type": "string"
					}
				},
				"required": [
					"Iteam_id",
					"IteamName",
					"Price",
					"password",
					"createdAt",
					"updatedAt"
				],
				"type": "object",
				"additionalProperties": false
			},
			"GetwaiterReqOder": {
				"properties": {
					"id": {
						"type": "string"
					}
				},
				"required": [
					"id"
				],
				"type": "object",
				"additionalProperties": false
			},
			"SaveReqgetmenuidORsaveOder": {
				"properties": {
					"menuId": {
						"type": "string"
					}
				},
				"required": [
					"menuId"
				],
				"type": "object",
				"additionalProperties": false
			},
			"waitersOder": {
				"$ref": "#/components/schemas/LeanDocument_this_"
			},
			"DeletewaiterReqOder": {
				"properties": {
					"id": {
						"type": "string"
					}
				},
				"required": [
					"id"
				],
				"type": "object",
				"additionalProperties": false
			}
		},
		"securitySchemes": {
			"api_keyjwt": {
				"type": "http",
				"name": "access_token",
				"in": "header",
				"scheme": "Bearer"
			},
			"api_key": {
				"type": "apiKey",
				"name": "access_token",
				"in": "query"
			}
		}
	},
	"info": {
		"title": "webapi",
		"version": "1.0.0",
		"license": {
			"name": "ISC"
		},
		"contact": {}
	},
	"openapi": "3.0.0",
	"paths": {
		"/admin/getWaitersById": {
			"post": {
				"operationId": "Getadmin",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/SaveUpdateResADMINwaiter"
								}
							}
						}
					}
				},
				"tags": [
					"Admin"
				],
				"security": [
					{
						"api_key": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/GetAdminwaiter"
							}
						}
					}
				}
			}
		},
		"/admin/AddNewWaiter": {
			"post": {
				"operationId": "Saveadmin",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/SaveUpdateResADMINwaiter"
								}
							}
						}
					}
				},
				"tags": [
					"Admin"
				],
				"security": [
					{
						"api_key": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/SaveReqADMINwaiter"
							}
						}
					}
				}
			}
		},
		"/admin/updateWaiterDetail": {
			"put": {
				"operationId": "UpdateAdmin",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/SaveUpdateResADMINwaiter"
								}
							}
						}
					}
				},
				"tags": [
					"Admin"
				],
				"security": [
					{
						"api_key": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UpdateReqADMINwaiter"
							}
						}
					}
				}
			}
		},
		"/admin/deletewaiter": {
			"delete": {
				"operationId": "Deletadmin",
				"responses": {
					"200": {
						"description": "Admin deleted",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ADMINwaiter"
								}
							}
						}
					}
				},
				"tags": [
					"Admin"
				],
				"security": [
					{
						"api_key": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/DeleteAdminwaiter"
							}
						}
					}
				}
			}
		},
		"/admin/getAdminwaiterList": {
			"get": {
				"operationId": "GetadminList",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/SaveUpdateResADMINwaiter"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"Admin"
				],
				"security": [
					{
						"api_key": []
					}
				],
				"parameters": []
			}
		},
		"/admin/AddNewMenu": {
			"post": {
				"operationId": "Savemenu",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/SaveUpdateResmenu"
								}
							}
						}
					}
				},
				"tags": [
					"Admin"
				],
				"security": [
					{
						"api_key": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/SaveReqmenu"
							}
						}
					}
				}
			}
		},
		"/admin/deleteMenu": {
			"delete": {
				"operationId": "DeletMenu",
				"responses": {
					"200": {
						"description": "Admin deleted",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/menuadmindoc"
								}
							}
						}
					}
				},
				"tags": [
					"Admin"
				],
				"security": [
					{
						"api_key": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/Deletemenu"
							}
						}
					}
				}
			}
		},
		"/admin/getMenuList": {
			"get": {
				"operationId": "GetMenusList",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/SaveUpdateResmenu"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"Admin"
				],
				"security": [
					{
						"api_key": []
					}
				],
				"parameters": []
			}
		},
		"/waiter/WaiterSearchOderById": {
			"post": {
				"operationId": "Getadmin",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/SaveUpdateReswaiterOder"
								}
							}
						}
					}
				},
				"tags": [
					"waiterOder_admin"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/GetwaiterReqOder"
							}
						}
					}
				}
			}
		},
		"/waiter/WaiterSaveoder": {
			"post": {
				"operationId": "Saveadmin",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/SaveUpdateReswaiterOder"
								}
							}
						}
					}
				},
				"tags": [
					"waiterOder_admin"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/SaveReqgetmenuidORsaveOder"
							}
						}
					}
				}
			}
		},
		"/waiter/waiterdeleteOder": {
			"delete": {
				"operationId": "Deletadmin",
				"responses": {
					"200": {
						"description": "waiter Oder deleted",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/waitersOder"
								}
							}
						}
					}
				},
				"tags": [
					"waiterOder_admin"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/DeletewaiterReqOder"
							}
						}
					}
				}
			}
		},
		"/waiter/Waitergetoderlist": {
			"get": {
				"operationId": "GetadminList",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/SaveUpdateReswaiterOder"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"waiterOder_admin"
				],
				"security": [],
				"parameters": []
			}
		},
		"/waiter/getMenuListw": {
			"get": {
				"operationId": "GetMenusListw",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/SaveUpdateResmenu"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"waiterOder_admin"
				],
				"security": [],
				"parameters": []
			}
		}
	},
	"servers": [
		{
			"url": "/"
		}
	]
}