# src/types/Request/menu.request.ts
export interface SaveReqmenu{
    IteamName: string;
    Price: number;
    
    }
    export interface UpdateReqmenu {
        Iteam_id:string;
        IteamName: string;
        Price: number;
       
       
    }
    export interface Getmenu {
      id: string
    }
    export interface Deletemenu {
      id: string
    }
  