# src/types/Responses/admin.waiter.Responses.ts
export interface SaveUpdateResADMINwaiter {
    _id:string;
    FirstName: string;
    LastName: string;
    Desgination: string;
    cell: string;
    JoinDate: string;
    adress: string;
    createdAt: string;
    updatedAt: string;
  }