# src/types/Responses/waiter.oder.responses.ts
export interface SaveUpdateReswaiterOder {
  Iteam_id:string;
  IteamName: string;
  Price: number;
    // Desgination: string;
    // cell: string;
    // email: string;
    password:string;
    createdAt: string;
    updatedAt: string;
  
  }